⭐✨🤩 huge performance improvement, remove lags.pauses from anim and checkKeyHeld the buttons or the pies were not enabled until the animation finish, which results in a laggy experience, fixed it, now buttons are always enabled, no need to wait for anim to finsh to activate a button.
Removed the blocking function "sleep" from checkKeyHeld, now key held or not is checked on a QTimer so on a seperte thread,
and the time is calculated using datetime.datetime.now, So, the program stops no where to check key held or to do anim stuff.

when menu was opened using hotkey TIME - CURRENT TIME in QTimer thread to calculate time elapsed.

Also added quick mouse gesture activate pies, so even if you quickly press and release hotkey (which acts like hotkey was
just pressed and not held) and move mouse quickly, program will detect it and not hold the pie menu => keyHeld = False